+++ Client/src/components/contanier/staff/Student/EditStudent/index.css
@@ -2,30 +2,25 @@
   margin: 50px;
 }
 
-.edit-student {
-  margin: 50px;
-  width: 100%;
-}
-
 .edit-student-title, 
 .edit-student-contanier, 
 .edit-student-title-border {
   color: #06468F;
 }
 
+.edit-student-title{
+  width: fit-content;
+  border-bottom: 2px solid;
+}
 .edit-student-contanier {
   width: 80%;
   display: flex;
   margin: 50px auto;
 }
 
-.edit-student-title-border {
-  border-bottom: 2px solid;
-}
-
 .edit-student-contanier-center {
   margin: auto 35px;
-  width: 4px;
+  width: 2px;
   background-color: #e0e0e0;
   height: 375px;
 }
@@ -36,16 +31,6 @@
   margin-top: 47px;
   height: 40px;
   width: 100%;
-}
-
-.danger-value {
-  border: 2px solid red;
-  box-shadow: 1px 0px 0px -1px red;
-}
-
-.success-value {
-  border: 2px solid green;
-  box-shadow: 1px 0px 0px -1px green;
 }
 
 .general-input {
