+++ Client/src/components/contanier/staff/Student/EditStudent/staticdata.js
@@ -1,0 +1,13 @@
+const leftSectionInfo = [
+  ['firstname', 'First name'],
+  ['lastname', 'Last name'],
+  ['level', 'Level'],
+  ['phonenumber', 'Phone number '],
+];
+const rightSectionInfo = [
+  ['mobilenumber', 'Mobile number'],
+  ['address', 'Address'],
+  ['password', 'Password']
+];
+
+export { leftSectionInfo, rightSectionInfo };
